etl_service/requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.7
pymongo==4.5.0